[ADD] add Views for chatroom, dm, friend. and rename template names
