+++ pong/app/javascript/srcs/templates/ChatRoomCollectionView.html
@@ -1,0 +1,1 @@
+ChatRoomCollectionView
