+++ pong/app/javascript/srcs/templates/DmRoomCollectionView.html
@@ -1,0 +1,1 @@
+DmRoomCollectionView
