+++ pong/app/javascript/srcs/templates/FriendCollectionView.html
@@ -1,0 +1,1 @@
+FriendCollectionView
