+++ pong/app/javascript/srcs/templates/LoginView.html
