+++ pong/app/javascript/srcs/templates/MainView.html
