+++ pong/app/javascript/srcs/templates/NavView.html
