+++ pong/app/javascript/srcs/templates/RootView.html
