+++ pong/app/javascript/srcs/templates/SideView.html
