+++ pong/app/javascript/srcs/views/ChatRoomCollectionView.js
@@ -1,0 +1,9 @@
+import common from '../common';
+import template from '../templates/ChatRoomCollectionView.html';
+
+const ChatRoomCollectionView = common.View.extend({
+  template,
+  onRender() {},
+});
+
+export default ChatRoomCollectionView;
