+++ pong/app/javascript/srcs/views/DmRoomCollectionView.js
@@ -1,0 +1,9 @@
+import common from '../common';
+import template from '../templates/DmRoomCollectionView.html';
+
+const DmRoomCollectionView = common.View.extend({
+  template,
+  onRender() {},
+});
+
+export default DmRoomCollectionView;
