+++ pong/app/javascript/srcs/views/FriendCollectionView.js
@@ -1,0 +1,9 @@
+import common from '../common';
+import template from '../templates/FriendCollectionView.html';
+
+const FriendCollectionView = common.View.extend({
+  template,
+  onRender() {},
+});
+
+export default FriendCollectionView;
