+++ pong/app/javascript/srcs/views/RootView.js
@@ -1,6 +1,6 @@
 import view from '.';
 import common from '../common';
-import template from '../templates/root.html';
+import template from '../templates/RootView.html';
 
 const RootView = common.View.extend({
   el: '#root',
